feat: implement custom `@@toStringTag`
There are several ways we can guard against a ConditionalExpression mutant that
may result in returning an incorrect Parrot object.

First solution would be to export all Parrot classes from the parrot module and
use `instanceof` to check whether the created object matches the expected
constructor. However, this would introduce a leaky abstraction as it would be
possible to use the classes directly instead of the `createParrot` factory
function.

Another solution is to use `Symbol.toStringTag` to create a custom
`@@toStringTag` symbol. This allows us to check for the correct object using
`Object.prototype.toString` without having to export the Parrot classes from
the parrot module.